# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Brand & Marketing OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /campaigns/{id}:
    get:
      operationId: getCampaign
      tags: [campaigns]
      summary: Get a campaign by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Campaign found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Campaign"
        "404":
          description: Campaign not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    CampaignDay:
      type: object
      properties:
        day:
          type: integer
        objective:
          type: string
        postConcept:
          type: string
        marketingAngle:
          type: string
        cta:
          type: string
      required:
        - day
        - objective
        - postConcept
        - marketingAngle
        - cta
    SocialPost:
      type: object
      properties:
        id:
          type: integer
        campaignId:
          type: integer
        day:
          type: integer
        caption:
          type: string
        hook:
          type: string
        cta:
          type: string
        hashtags:
          type: array
          items:
            type: string
        imagePrompt:
          type: string
        imageUrl:
          type: ["string", "null"]
          description: Base64 data URL of the AI-generated image
        platform:
          type: string
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - campaignId
        - day
        - caption
        - hook
        - cta
        - hashtags
        - imagePrompt
        - platform
        - createdAt
        - updatedAt
...
    Campaign:
      type: object
      properties:
        id:
          type: integer
        brandId:
          type: integer
        title:
          type: string
        strategy:
          type: string
        days:
          type: array
          items:
            $ref: "#/components/schemas/CampaignDay"
        posts:
          type: array
          items:
            $ref: "#/components/schemas/SocialPost"
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - brandId
        - title
        - strategy
        - days
        - posts
        - createdAt
        - updatedAt
...